+++ NEWS.md
@@ -52,6 +52,9 @@
  * Add support for domain-specific CRAN package **partykit**, e.g. `cf
    <- cforest(dist ~ speed, data = cars) |> futurize()`.
 
+ * Add support for domain-specific CRAN package **rmgarch**,
+   e.g. `fit <- dccfit(spec, data = returns) |> futurize()`.
+
  * Add support for domain-specific CRAN package **Rsolnp**,
    e.g. `res <- gosolnp(fun, LB, UB, n.restarts = 5) |> futurize()`.